헤더 / 구현 분리
