+++ .gitignore
@@ -72,3 +72,6 @@
 /Debug
 /.vs/OOPTeam15Project4/FileContentIndex
 /.vs/OOPTeam15Project4/v17/ipch/AutoPCH/9da6bb472a30c1ad/MAIN.ipch
+/x64/Debug/oop rhythm.Build.CppClean.log
+/x64/Debug
+/oop rhythm.vcxproj
